removed admin access function and merged it into verify user function

+++ src/blueprints/auth_bp.py
@@ -103,35 +103,20 @@ def delete_user(user_id):
             <user_id> (int)    
     """
     admin_delete_user = verify_user()
+    if not admin_delete_user:
+        return {"forbidden": "You must be logged in or registered"}, 403
     # verify user
     if admin_delete_user:
         if admin_delete_user.id == user_id and not admin_delete_user.is_admin:
             db.session.delete(admin_delete_user)
             db.session.commit()
             return {'deleted': 'user successfully deleted'}
-        admin_user = admin_access()
-        if admin_user:
-            db.session.delete(admin_delete_user)
-            db.session.commit()
-            return {'admin_deleted': 'user successfully deleted'}
+        if not admin_delete_user.is_admin:
+            return {"unauthorized" : "Admin access only"}, 401
+        db.session.delete(admin_delete_user)
+        db.session.commit()
+        return {'admin_deleted': 'user successfully deleted'}
     return {"invalid_user" : "User not found"}, 404
-
-
-def admin_access():
-    """
-    Admin Access Function
-
-    Allows only the admin user to access the route 
-    this function is called in. Function is placed in the view function
-    """
-    # get the user's identity
-    user_id = get_jwt_identity()
-    # query the database and check if the user is_admin
-    stmt = db.select(User).filter_by(id=user_id)
-    user = db.session.scalar(stmt)
-    if not user.is_admin:
-        abort(401, description='admin access only')
-    return user
 
 def verify_user_car(car_id):
     """
@@ -154,6 +139,4 @@ def verify_user():
     """
     stmt = db.select(User).filter_by(id=get_jwt_identity())
     user = db.session.scalar(stmt)
-    if user:
-        return user
-    abort(401, "You must be logged in or registered")
+    return user
